refactor veto logic
